fix: tests e2e
