+++ package.json
@@ -44,6 +44,7 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.37.0",
+    "@faker-js/faker": "^10.1.0",
     "@types/node": "^24.5.1",
     "@typescript-eslint/eslint-plugin": "^8.46.1",
     "@typescript-eslint/parser": "^8.46.1",
